refactor: move as classes para as pastas específicas.
