+++ src/main/groovy/com/mini/asaas/base/BaseEntity.groovy
@@ -1,14 +1,13 @@
-package core.entities
+package com.mini.asaas.base
 
 import java.time.LocalDateTime
 
-/**
- * <p>Classe abstrata que representa uma entidade base.</p>
- * <p>Essa classe contém os atributos comuns a todas as entidades.</p>
- */
 abstract class BaseEntity {
+
     LocalDateTime dateCreated
+
     LocalDateTime lastUpdated
+
     Boolean deleted = false
 
     static mapping = {
